Add initial structure for InterestRate module with DTO, entity, mapper, repository, service, and service implementation

+++ server/src/main/java/com/lankacapital/server/dtos/LoanCreateDto.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 public class LoanCreateDto {
     private String fileNumber;
     private BigDecimal loanAmount;
-    private Double interestRate;
+    private Integer interestRate;
     private BigDecimal documentCharge;
     private Double numberOfInstallments;
     private Long employeeId;
